refactor: rename modal component files
