+++ src/components/modal/modal-component.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, jest } from '@jest/globals'
 import React from 'react'
 import { render, screen, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import { Modal, ModalActions, ModalBody } from './Modal.js'
+import { Modal, ModalActions, ModalBody } from './modal-component.js'
 
 describe('Modal', () => {
   describe('Modal component', () => {
